\name{check.start}
\alias{check.start}
\title{
  Check starting point
}
\description{
  Program used to check whether the starting point is a good starting point. 
}
\usage{
check.start(clust)
}
\arguments{  
  \item{clust}{
    The initial starting point for the Markov chain  
  }
}
\value{
  The logic value describe whether it is a good starting point.
}
\section{References}{ 
   Ma, P., Castillo-Davis, C., Zhong, W., and Liu, J. S. (2006)
   A data-driven clustering method for time course gene expression data,  
   Nucleic Acids Research, 34 (4), 1261-1269. 
}
\seealso{
...
}
\examples{
}
\keyword{cluster}
