\name{testdata}
\docType{data}
\alias{testdata}
\title{Time course simulated data}
\description{
	Simulated functional clusterd data with each row represents a subject and each column representing a time point. Data are generated from 4 functions with 30 curve from \eqn{f1=-exp(tt)/1000}, 40 curves from \eqn{f2=tan(tt)}, 50 curves from \eqn{f3=5*(tt-4)^2/max((tt-4)^2)}, 30 curves from \eqn{f4=cos(tt)}}
\usage{data(testdata)}
\format{a matrix with 150 rows and 10 columns}
\source{simulated data}
\keyword{datasets}
